Moved parse_timestamps.py to g2lib directory. Changed default path to expanduser in readevents
